refactor(workflow): centralize task execution types and error handling
