+++ src/qdash/workflow/engine/__init__.py
@@ -109,7 +109,7 @@ from qdash.workflow.engine.scheduler.one_qubit_scheduler import (
 
 # Task components
 from qdash.workflow.engine.task.context import TaskContext
-from qdash.workflow.engine.task.executor import TaskExecutionError, TaskExecutor
+from qdash.workflow.engine.task.executor import TaskExecutor
 from qdash.workflow.engine.task.history_recorder import TaskHistoryRecorder
 from qdash.workflow.engine.task.result_processor import (
     FidelityValidationError,
@@ -117,6 +117,7 @@ from qdash.workflow.engine.task.result_processor import (
     TaskResultProcessor,
 )
 from qdash.workflow.engine.task.state_manager import TaskStateManager
+from qdash.workflow.engine.task.types import TaskExecutionError
 
 __all__ = [
     # Orchestration
